Consolidate to a single file and export validateMatches
The library is simple enough, and this makes consuming easier
if this library is being bundled into a WebWorker as a single
script dependency when using CJS or ESM builds.
