+++ test/index.test.ts
@@ -3,7 +3,7 @@ import * as Exports from '../src';
 describe('index', () => {
   it('exports the right stuff', () => {
     expect(Exports.calculate).toBeDefined();
-    expect(Exports.calculateSync).toBeDefined();
-    expect(Exports.DerangementError).toBeDefined();
+    expect(Exports.validateMatches).toBeDefined();
+    expect(Object.keys(Exports)).toEqual(['validateMatches', 'calculate']);
   });
 });
